refactor: logical grouping of files in lib and improved separation of concerns
