+++ src/app/api/course-list/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { connectToDatabase } from "@/lib/mongoose";
+import { connectToDatabase } from "@/lib/database/mongoose";
 import { Course } from "@/db/course";
 
 export const dynamic = "force-dynamic";
