+++ src/app/api/papers/count/route.ts
@@ -1,5 +1,5 @@
 import { NextResponse } from "next/server";
-import { connectToDatabase } from "@/lib/mongoose";
+import { connectToDatabase } from "@/lib/database/mongoose";
 import CourseCount from "@/db/course";
 
 export const dynamic = "force-dynamic";
