+++ src/app/api/papers/route.ts
@@ -1,7 +1,9 @@
 import { NextResponse, type NextRequest } from "next/server";
-import { connectToDatabase } from "@/lib/mongoose";
+import { connectToDatabase } from "@/lib/database/mongoose";
 import Paper from "@/db/papers";
 import { type IPaper } from "@/interface";
+import { escapeRegExp } from "@/lib/utils/regex";
+import { extractUniqueValues } from "@/lib/utils/paper-aggregation";
 
 export const dynamic = "force-dynamic";
 
@@ -10,10 +12,6 @@ export async function GET(req: NextRequest) {
     await connectToDatabase();
     const url = req.nextUrl.searchParams;
     const subject = url.get("subject");
-    const escapeRegExp = (text: string) => {
-      return text.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-    };
-    const escapedSubject = escapeRegExp(subject ?? "");
 
     if (!subject) {
       return NextResponse.json(
@@ -22,42 +20,17 @@ export async function GET(req: NextRequest) {
       );
     }
 
+    const escapedSubject = escapeRegExp(subject);
     const papers: IPaper[] = await Paper.find({
       subject: { $regex: new RegExp(`${escapedSubject}`, "i") },
     });
 
-    if (papers.length === 0) {
-      return NextResponse.json(
-        {
-          papers,
-          unique_years: [],
-          unique_slots: [],
-          unique_exams: [],
-          unique_campuses: [],
-          unique_semesters: [],
-        },
-        { status: 200 },
-      );
-    }
-
-    const unique_years = Array.from(new Set(papers.map((paper) => paper.year)));
-    const unique_slots = Array.from(new Set(papers.map((paper) => paper.slot)));
-    const unique_exams = Array.from(new Set(papers.map((paper) => paper.exam)));
-    const unique_campuses = Array.from(
-      new Set(papers.map((paper) => paper.campus)),
-    );
-    const unique_semesters = Array.from(
-      new Set(papers.map((paper) => paper.semester)),
-    );
+    const uniqueValues = extractUniqueValues(papers);
 
     return NextResponse.json(
       {
         papers,
-        unique_years,
-        unique_slots,
-        unique_exams,
-        unique_campuses,
-        unique_semesters,
+        ...uniqueValues,
       },
       { status: 200 },
     );
