+++ src/app/api/upcoming-papers/route.ts
@@ -1,7 +1,8 @@
 import { NextResponse } from "next/server";
-import { connectToDatabase } from "@/lib/mongoose";
+import { connectToDatabase } from "@/lib/database/mongoose";
 import UpcomingSlot from "@/db/upcoming-slot";
 import UpcomingSubject from "@/db/upcoming-paper";
+import { calculateCorrespondingSlots } from "@/lib/utils/slot-calculation";
 
 export const dynamic = "force-dynamic";
 
@@ -10,6 +11,7 @@ export async function GET() {
     await connectToDatabase();
     const upcomingSlot = await UpcomingSlot.find();
     const slot = upcomingSlot[0]?.slot;
+    
     if (!slot) {
       return NextResponse.json(
         {
@@ -18,16 +20,12 @@ export async function GET() {
         { status: 404 },
       );
     }
-    const nextSlot = String.fromCharCode(slot.charCodeAt(0) + 1);
-    const correspondingSlots = [
-      slot + "1",
-      slot + "2",
-      nextSlot + "1",
-      nextSlot + "2",
-    ];
+
+    const correspondingSlots = calculateCorrespondingSlots(slot);
     const selectedSubjects = await UpcomingSubject.find({
       slots: { $in: correspondingSlots },
     });
+    
     if (selectedSubjects.length === 0) {
       return NextResponse.json(
         {
