+++ src/lib/storage/gcp.ts
@@ -3,7 +3,6 @@ import { Storage, StorageOptions } from "@google-cloud/storage";
 interface GCPCredentials {
   type: string;
   project_id: string;
-  private_key_id: string;
   private_key: string;
   client_email: string;
   client_id: string;
